% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Boundaryconstruction.R
\name{Boundaryconstruction}
\alias{Boundaryconstruction}
\title{Boundaryconstruction}
\usage{
Boundaryconstruction(Stopbound.inf = Stopbound.inf, ns = ns)
}
\arguments{
\item{Stopbound.inf}{The list of stop boundary information for more see \code{\link{Stopboundinf}}}

\item{ns}{A vector of accumulated number of patient at each stage}
}
\value{
A list of the futility boundary and the efficacy boundary
}
\description{
This function constructs the stopping boundary based on input information
}
\examples{
 Stopbound.inf=list(Stop.type="Early-Pocock",Boundary.type="Symmetric",cutoff=c(0.9928,0.0072))
 ns=c(60,120,180,240,300)
 Boundaryconstruction(Stopbound.inf, ns)
}
\author{
Ziyan Wang
}
