% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R
\name{conjuncativepower_or_FWER}
\alias{conjuncativepower_or_FWER}
\title{conjuncativepower_or_FWER}
\usage{
conjuncativepower_or_FWER(res, scenario, test.type)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}

\item{scenario}{The true scenario used to generate the res list}

\item{test.type}{The indicator of whether using one side or two side testing.
Please make sure that the input test.type does not conflicts to the data. Otherwise the conjunctive power calculation is wrong}
}
\value{
Family wise error rate or Conjunctive power
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate the Family wise error rate or Conjunctive power
}
\examples{
\dontrun{conjuncativepower_or_FWER(res)}
}
