% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_document.R
\docType{data}
\name{dataloginformd}
\alias{dataloginformd}
\title{Cutoff screening example: the details of grid}
\format{
A data frame with 24 rows and 3 variables:
\describe{
\item{\code{tpIE}}{FWER}
\item{\code{cutoff}}{Cutoff value}
\item{\code{cutoff2}}{Square of cutoff value}
}
}
\usage{
dataloginformd
}
\description{
Details of grid including famliy wise error rate of a cutoff value, the cutoff value and the square of cutoff value for modelling and prediction
}
\keyword{datasets}
