% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R
\name{FWER_disconjunctivepowerfunc}
\alias{FWER_disconjunctivepowerfunc}
\title{FWER_disconjunctivepowerfunc}
\usage{
FWER_disconjunctivepowerfunc(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
Family wise error rate or disconjunctive power
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate the Family wise error rate or disconjunctive power
}
\examples{
\dontrun{FWER_disconjunctivepowerfunc(res)}
}
