% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_listoffunctionforplotgeneration.R
\name{intbias}
\alias{intbias}
\title{intbias}
\usage{
intbias(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
A matrix of mean treatment - stage interaction effect bias
}
\description{
This function estimates the mean bias of treatment - stage interaction effect
}
\examples{
\dontrun{intbias(res)}
}
