% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R
\name{Meanfunc}
\alias{Meanfunc}
\title{Meanfunc}
\usage{
Meanfunc(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
Mean treatment effect estimates of each treatment arm
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate mean treatment effect estimate.
}
\examples{
\dontrun{Meanfunc(res)}
}
