% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R
\name{Nfunc}
\alias{Nfunc}
\title{Nfunc}
\usage{
Nfunc(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
The mean estimate of total number of patients allocated to each arm
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate mean estimate of total number of patients allocated to each arm
}
\examples{
\dontrun{Nfunc(res)}

}
