% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_document.R
\docType{data}
\name{OPC_null}
\alias{OPC_null}
\title{Operation characteristic table for null scenario}
\format{
A data frame with 3 rows and 16 variables:
\describe{
\item{\code{Type.I.Error.or.Power}}{Family wise error rate}
\item{\code{Bias.1}}{Treatment effect bias for treatment 1}
\item{\code{Bias.2}}{Treatment effect bias for treatment 2}
\item{\code{Bias.3}}{Treatment effect bias for treatment 3}
\item{\code{rMSE.1}}{Rooted mean squared error for treatment 1}
\item{\code{rMSE.2}}{Rooted mean squared error for treatment 2}
\item{\code{rMSE.3}}{Rooted mean squared error for treatment 3}
\item{\code{N.per.arm.1}}{Mean total number of patient allocated to control}
\item{\code{N.per.arm.2}}{Mean total number of patient allocated to treatment 1}
\item{\code{N.per.arm.3}}{Mean total number of patient allocated to treatment 2}
\item{\code{N.per.arm.4}}{Mean total number of patient allocated to treatment 3}
\item{\code{Survive.per.arm.1}}{Mean total number of patient allocated to control}
\item{\code{Survive.per.arm.2}}{Mean total number of patient survived when using treatment 1}
\item{\code{Survive.per.arm.3}}{Mean total number of patient survived when usin treatment 2}
\item{\code{Survive.per.arm.4}}{Mean total number of patient survived when usin treatment 3}
\item{\code{N}}{Mean total number of patient in a trial}
}
}
\usage{
OPC_null
}
\description{
Operation characteristic table for null scenario using main and main + continuousstage model. The main effect model was run for a null scenario with and without time trend.
The time trend pattern is step. The strength of time trend is 0.1 equally for all arm. The effect of time trend on true response probability is multiplicative.
}
\keyword{datasets}
