% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_document.R
\docType{data}
\name{OPC_Trial.simulation}
\alias{OPC_Trial.simulation}
\title{Operation characteristic table for Trial.simulation() null scenario}
\format{
A data frame with 1 rows and 8 variables:
\describe{
\item{\code{Type.I.Error.or.Power}}{Power}
\item{\code{Bias}}{Treatment effect bias for treatment 1}
\item{\code{rMSE}}{Rooted mean squared error for treatment effect 1}
\item{\code{N.per.arm.1}}{Mean total number of patient allocated to control}
\item{\code{N.per.arm.2}}{Mean total number of patient allocated to treatment 1}
\item{\code{Survive.per.arm.1}}{Mean total number of patient allocated to control}
\item{\code{Survive.per.arm.2}}{Mean total number of patient survived when using treatment 1}
\item{\code{N}}{Mean total number of patient in a trial}
}
}
\usage{
OPC_Trial.simulation
}
\description{
Operation characteristic table for null scenario using main model.
}
\keyword{datasets}
