% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_document.R
\docType{data}
\name{optimdata_asy}
\alias{optimdata_asy}
\title{A list of data from Gaussian process and trial simulation for asymmetric cutoff screening.}
\format{
A list with four element:
\describe{
\item{\code{next.cutoff}}{The cutoff value for the next evaluation}
\item{\code{prediction}}{A list of values from Gaussian process model}
\item{\code{ESS}}{A two column twenty five rows matrix with the effective sample size for each cutoff pair under both null (first column) and alternative (second column) scenario}
\item{\code{testeddata}}{A data frame containing each tested cutoff pair (column two and three for efficacy and futility, respectively), their FWER under null (column one) and conjunctive power under alternative (column four)}
}
}
\usage{
optimdata_asy
}
\description{
A list of data from Gaussian process and trial simulation for asymmetric cutoff screening.
}
\keyword{datasets}
