% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Outputmatrixinitialization.R
\name{OutputStats.initialising}
\alias{OutputStats.initialising}
\title{OutputStats.initialising}
\usage{
OutputStats.initialising(variable.inf, reg.inf, ns, K)
}
\arguments{
\item{variable.inf}{The parameter information in the model}

\item{reg.inf}{The model information. For the fixed effect model, the input of reg.inf can be main, main + stage_continuous, main * stage_continuous, main + stage_discrete,
main * stage_discrete.
For the mixed effect model, the reg.inf is invalid.}

\item{ns}{A vector of accumulated number of patient at each stage}

\item{K}{Total number of arm including control}
}
\value{
The empty output matrix including different evaluation metrics.
}
\description{
This function initializes the output matrix including all evaluation metrics based on input information
}
\examples{
OutputStats.initialising(
    variable.inf = "Fixeffect",
    reg.inf = "main",
    ns = c(15, 30, 45, 60, 75),
    K = 2)
}
\author{
Ziyan Wang
}
