% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R
\name{perHtypeIerror_marginalpowerfunc}
\alias{perHtypeIerror_marginalpowerfunc}
\title{perHtypeIerror_powerfunc}
\usage{
perHtypeIerror_marginalpowerfunc(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
"per-hypothesis" Type I error rate or power (marginal power) for each treatment - control comparison
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate the error rate
}
\examples{
\dontrun{perHtypeIerror_powerfunc(res)}
}
\author{
Ziyan Wang
}
