% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R
\name{perHtypeIerror_powerfunc}
\alias{perHtypeIerror_powerfunc}
\title{perHtypeIerror_powerfunc}
\usage{
perHtypeIerror_powerfunc(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
Type I error rate or power of each treatment - control comparison
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate the error rate
}
\examples{
\dontrun{perHtypeIerror_powerfunc(res)}
}
\author{
Ziyan Wang
}
