% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_document.R
\docType{data}
\name{predictedtpIEinformd}
\alias{predictedtpIEinformd}
\title{Cutoff screening example: the predicted value from quadratic model}
\format{
A data frame with 1001 rows and 1 variables:
\describe{
\item{\code{predictedtpIEinformd}}{The predicted FWER value of a large grid}
}
}
\usage{
predictedtpIEinformd
}
\description{
The predicted value from quadratic model for famliy wise error rate vs cutoff value plotting
}
\keyword{datasets}
