% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Randomisationinformationcheck.R
\name{Randomisation.inf}
\alias{Randomisation.inf}
\title{Randomisation.inf}
\usage{
Randomisation.inf(
  Random.inf = list(Fixratio = FALSE, Fixratiocontrol = NA, BARmethod = "Thall",
    Thall.tuning.inf = list(tuningparameter = "Fixed", fixvalue = 1), Trippa.tuning.inf =
    list(a = 10, b = 0.75))
)
}
\arguments{
\item{Random.inf}{A list of adaptive randomisation information.
'Fixratio' a indicator of whether the randomisation process uses fix ratio. Default is FALSE.
'Fixratiocontrol' the numerical value indicating the randomisation weight of the control arm compared to the treatment arms. Default is NA for Fixratio = FALSE.
'BARmethod' the Bayesian adaptive randomisation type. Default is "Thall" indicating the use of Thall's approach in the randomisation process. The other value is 'Trippa'.
'Thall.tuning.inf' the list of tuning parameter for Thall's approach including 'tuningparameter' (Default is "Fixed" indicating that the tuning paramter is fixed for all stages) and 'fixvalue' (Default is 1).}
}
\value{
A list of input randomisation information
}
\description{
This function checks the validity of the randomisation information input
}
\examples{
Randomisation.inf(Random.inf = list(
Fixratio = FALSE,
Fixratiocontrol = NA,
BARmethod = "Thall",
Thall.tuning.inf = list(tuningparameter = "Fixed",  fixvalue = 1),
Trippa.tuning.inf = list(a = NA, b = NA)
))
}
\author{
Ziyan Wang
}
