% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafile_document.R
\docType{data}
\name{recommandloginformd}
\alias{recommandloginformd}
\title{Cutoff screening example: the recommended grid value at each time point}
\format{
A data frame with 20 rows and 1 variables:
\describe{
\item{\code{recommandloginformd}}{The cutoff value at each time point}
}
}
\usage{
recommandloginformd
}
\description{
he recommended grid value at each time point. There are 20 cutoff value explored
}
\keyword{datasets}
