% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Stansamplesummary_Fixeffect.R
\name{resultrtostats}
\alias{resultrtostats}
\title{resultrtostats}
\usage{
resultrtostats(
  trteff = NA,
  treatmentindex = NA,
  armleft,
  K,
  group,
  reg.inf,
  fit,
  ns
)
}
\arguments{
\item{trteff}{Stan posterior samples of treatment effect sample distribution}

\item{treatmentindex}{A vector of treatment index at the beginning of a trial}

\item{armleft}{The number of treatment left in the platform (>2)}

\item{K}{Total number of arms at the beginning}

\item{group}{The current stage}

\item{reg.inf}{The information of how much accumulated information will be used}

\item{fit}{The stan output}

\item{ns}{A vector of accumulated number of patient at each stage}
}
\value{
A list of stan result inference
stats1: A vector of posterior probability for all treatment arms
including dropped and active treatment arm
stats4: The mean treatment effect estimate of each treatment compared to control
stats5: The variance of treatment effect estimate of each treatment compared to control
post.prob.btcontrol: a vector including Posterior probability
of each active treatment arm better than control
}
\description{
This is an inner function of the function \code{\link{resultstantoRfunc}}
}
\examples{
\dontrun{resultrtostats(trteff = NA, treatmentindex = NA, armleft, K, group, reg.inf, fit, ns)}
}
\author{
Ziyan Wang
}
