% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation_Stansamplesummarytolist_Mixeffect.R
\name{resultstantoRfunc.rand}
\alias{resultstantoRfunc.rand}
\title{resultstantoRfunc.rand}
\usage{
resultstantoRfunc.rand(group, fit, armleft, treatmentindex, K, ns)
}
\arguments{
\item{group}{The current stage}

\item{fit}{The stan output}

\item{armleft}{The number of treatment left in the platform (>2)}

\item{treatmentindex}{A vector of treatment index at the beginning of a trial}

\item{K}{Total number of arms at the beginning}

\item{ns}{A vector of accumulated number of patient at each stage}
}
\value{
A list of stan result inference
stats1: A vector of posterior probability for all treatment arms
including dropped and active treatment arm
stats4: The mean treatment effect estimate of each treatment compared to control
stats5: The variance of treatment effect estimate of each treatment compared to control
post.prob.btcontrol: a vector including Posterior probability
of each active treatment arm better than control
stats6: A vector of mean stage (time trend) effect estimate
stats7: A vector of mean treatment - stage (time trend) interaction effect estimate
sampefftotal: The posterior samples of response probability
of each active arm on logit scale.
This can be transformed to probit scale by using inv.logit() function.
}
\description{
This function summarise the mix effect stan output data to and transform them to be readable.
}
\examples{
\dontrun{resultstantoRfunc.rand(group, fit, armleft, treatmentindex, K, ns)}
}
\author{
Ziyan Wang
}
