% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R,
%   R/Plot_listoffunctionforplotgeneration.R
\name{Sperarmfunc}
\alias{Sperarmfunc}
\title{Sperarmfunc}
\usage{
Sperarmfunc(res)

Sperarmfunc(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
The mean total number of survived patients of each arm

The mean total number of survived patients of each arm
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate mean total number of survived patients of each arm

This function reads in the output matrix of a number of trial replicates to calculate mean total number of survived patients of each arm
}
\examples{
\dontrun{Sperarmfunc(res)}
\dontrun{Sperarmfunc(res)}
}
