% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_listoffunctionforplotgeneration.R
\name{trtbias}
\alias{trtbias}
\title{trtbias}
\usage{
trtbias(res, trueeffect)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}

\item{trueeffect}{A vector of true treatment effect in each scenario}
}
\value{
A matrix of mean treatment effect bias
}
\description{
This function estimates the mean bias of treatment effect
}
\examples{
\dontrun{trtbias(res, trueeffect)}
}
