% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_listoffunctionforplotgeneration.R
\name{trteffect}
\alias{trteffect}
\title{trteffect}
\usage{
trteffect(res, trueeff)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}

\item{trueeff}{A vector of true treatment effect in each scenario}
}
\value{
A vector of mean treatment effect bias and its rooted mean squared error
}
\description{
This function estimates the mean treatment effect bias and its rooted mean squared error
}
\examples{
\dontrun{trteffect(res, trueeff)}
}
