% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis_listofanalysisfunction.R
\name{varfunc}
\alias{varfunc}
\title{varfunc}
\usage{
varfunc(res)
}
\arguments{
\item{res}{A list of output matrix of a number of trial replicates}
}
\value{
The variance of Treatment effect estimates of each treatment arm
}
\description{
This function reads in the output matrix of a number of trial replicates to calculate the variance of treatment effect estimate.
}
\examples{
\dontrun{varfunc(res)}
}
